Corrected and updated translation with cleanup.
Merged pattern shape for variable and non-binding to 'arbitrary'.
Added abstraction2 to js_term as it is needed for correct translation
from core syntax's abstraction2.
Different translation of LetRec (is now a sequence).
Correct translation of abstraction and abstraction2 - wraps the term
to a 'let var = projection'.. also created necesarry match variables.
Correct projection for PVariant - a new type - VariantProj.