Add translation of MATCH clauses.
Add helper methods for pattern matching to `jsPervasives.js`.
Add some methods to `jsPervasives.eff`.

Removed `js_term` as the first term of `Match` as it is not needed.
Add an optional parameter to `of_abstraction` function to pass in a match variable.
Labels are printed as un-safe.
Extract printing of contants.